+++ CHANGELOG.md
@@ -8,7 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Fixed
 
 * Some `Konva.Transformer` fixes
-
+* Typescript fixes
+* `stage.toDataURL()` fixes when it has hidden layers
 
 ## [2.1.7][2018-07-03]